Deleting Points/Rebounds for concision and added more comments
I deleted the points and rebounds views, templates, and urls. I did this so that all the data could be stored in the index.html file where the user could see the points per game, rebounds per game, 2 pt %, 3 pt %, and games played. (I had to delete a lot of code which made this process more streamlined but makes the views look relatively simple.)